Moving old unification utils
