+++ lisa-utils/src/main/scala/lisa/utils/unification/UnificationUtils2.scala
@@ -5,599 +5,587 @@ import lisa.utils.KernelHelpers.{_, given}
 import lisa.utils.unification.FirstOrderUnifier.*
 
 /**
- * General utilities for unification, substitution, and rewriting
+ * General utilities for unification, substitution, and rewrites
  */
 object UnificationUtils2 {
 
-  extension [A](seq: Seq[A]) {
-
-    /**
-     * Seq.collectFirst, but for a function returning an Option. Evaluates the
-     * function only once per argument. Returns when the first non-`None` value
-     * is found.
-     *
-     * @param T output type under option
-     * @param f the function to evaluate
-     */
-    def getFirst[T](f: A => Option[T]): Option[T] = {
-      var res: Option[T] = None
-      val iter = seq.iterator
-      while (res.isEmpty && iter.hasNext) {
-        res = f(iter.next())
-      }
-      res
-    }
-
-  }
+  type LambdaTermFormulaFormula = (Seq[VariableFormulaLabel], Seq[VariableLabel], Formula)
 
   /**
-   * All the information required for performing rewrites.
-   */
-  case class RewriteContext(
-      freeFormulaRules: Seq[(Formula, Formula)] = Seq.empty,
-      freeTermRules: Seq[(Term, Term)] = Seq.empty,
-      confinedFormulaRules: Seq[(Formula, Formula)] = Seq.empty,
-      confinedTermRules: Seq[(Term, Term)] = Seq.empty,
-      takenFormulaVars: Set[VariableFormulaLabel] = Set.empty,
-      takenTermVars: Set[VariableLabel] = Set.empty
-  ) {
-    private var lastID: Identifier = freshId((takenFormulaVars ++ takenTermVars).map(_.id), "@@rewriteVar@@")
-
-    /**
-     * Generates a fresh identifier with an internal label `__rewriteVar__`.
-     * Mutates state.
-     *
-     * @return fresh identifier
-     */
-    def freshIdentifier = {
-      lastID = freshId(Seq(lastID), "@@rewriteVar@@")
-      lastID
-    }
-
-    def isFreeVariable(v: VariableLabel) = !takenTermVars.contains(v)
-    def isFreeVariable(v: VariableFormulaLabel) = !takenFormulaVars.contains(v)
-
-    /**
-     * Update the last generated fresh ID to that of another context if it is
-     * larger, otherwise retain the previous value. Mutates state.
-     *
-     * @param other another context
-     */
-    def updateTo(other: RewriteContext) =
-      lastID = if (other.lastID.no > lastID.no) other.lastID else lastID
-  }
-
-  object RewriteContext {
-    def empty = RewriteContext()
-  }
-
-  // substitutions
-
-  type TermSubstitution = Map[VariableLabel, Term]
-  val TermSubstitution = Map // don't abuse pls O.o
-
-  type FormulaSubstitution = Map[VariableFormulaLabel, Formula]
-  val FormulaSubstitution = Map
-
-  /**
-   * Performs first-order matching for two terms. Returns a (most-general)
-   * substitution from variables to terms such that `first` substituted is equal
-   * to `second`, if one exists. Uses [[matchTermRecursive]] as the actual
-   * implementation.
+   * Helper function for [[canReachOneStepTerm]]
    *
-   * @param reference the reference term
-   * @param template the term to match
-   * @param takenVariables any variables in the template which cannot be
-   * substituted, i.e., treated as constant
-   * @return substitution (Option) from variables to terms. `None` iff a
-   * substitution does not exist.
+   * @param first term
+   * @param second term
+   * @param subst list of possible substitutions, with assigned identifiers
    */
-  def matchTerm(reference: Term, template: Term, takenVariables: Iterator[VariableLabel] = Iterator.empty): Option[TermSubstitution] = {
-    val context = RewriteContext(takenTermVars = takenVariables.toSet)
-    matchTermRecursive(using context)(reference, template, TermSubstitution.empty)
+  def canReachOneStepTerm2(first: Term, second: Term, subst: Seq[((Term, Term), Identifier)]): Option[Term] = {
+    lazy val validSubst = subst.find { case ((l, r), _) => isSameTerm(first, l) && isSameTerm(second, r) }
+
+    if (isSameTerm(first, second)) Some(first)
+    else if (validSubst.isDefined) Some(VariableLabel(validSubst.get._2))
+    else if (first.label == second.label && first.args.length == second.args.length) {
+      val argCan = (first.args zip second.args).map { case (f, s) => canReachOneStepTerm2(f, s, subst) }
+
+      if (argCan.exists(_.isEmpty)) None
+      else Some(Term(first.label, argCan.map(_.get)))
+    } else None
   }
 
   /**
-   * Implementation for matching terms. See [[matchTerm]] for the interface.
+   * Helper function for [[canReachOneStepTermFormula]]
    *
-   * @param context all information about restricted variables and fresh name
-   * generation state
-   * @param reference the reference terms
-   * @param template the terms to match
-   * @param substitution currently accumulated susbtitutions to variables
-   * @return substitution (Option) from variables to terms. `None` if a
-   * substitution does not exist.
+   * @param first formula
+   * @param second formula
+   * @param subst list of term substitutions, with assigned identifiers
+   * @param takenIds list of taken identifiers for instantiation
    */
-  private def matchTermRecursive(using context: RewriteContext)(reference: Term, template: Term, substitution: TermSubstitution): Option[TermSubstitution] =
-    if (reference == template)
-      Some(substitution)
+  def canReachOneStepTerm2(first: Formula, second: Formula, subst: Seq[((Term, Term), Identifier)], takenIds: Set[Identifier]): Option[Formula] = {
+    if (isSame(first, second)) Some(first)
+    else if (first.label != second.label) None
     else
-      template.label match {
-        case v @ VariableLabel(id) if context.isFreeVariable(v) =>
-          // different label but substitutable or already correctly set
-          if (reference.label != template.label && reference == substitution.getOrElse(v, reference)) Some(substitution + (v -> reference))
-          // same and not already substituted to something else
-          else if (reference.label == template.label && reference == substitution.getOrElse(v, reference)) Some(substitution)
-          // unsat
-          else None
-        // {Constant, Schematic} FunctionLabel
-        case _ =>
-          if (reference.label != template.label) None
-          else
-            (reference.args zip template.args).foldLeft(Option(substitution)) {
-              case (Some(subs), (r, t)) => matchTermRecursive(r, t, subs)
-              case (None, _) => None
+      first match {
+        case PredicateFormula(l1, arg1: Seq[Term]) =>
+          second match {
+            case PredicateFormula(l2, arg2: Seq[Term]) => {
+              val argCan = (arg1 zip arg2).map { case (f, s) => canReachOneStepTerm2(f, s, subst) }
+
+              if (argCan.exists(_.isEmpty)) None
+              else Some(PredicateFormula(l1, argCan.map(_.get)))
             }
-      }
+            case _ => None
+          }
+        case ConnectorFormula(l1, arg1) => {
+          second match {
+            case ConnectorFormula(l2, arg2) => {
+              val argCan = (arg1 zip arg2).map { case (f, s) => canReachOneStepTerm2(f, s, subst, takenIds) }
 
-  /**
-   * Performs first-order matching for two formulas. Returns a (most-general)
-   * substitution from variables to terms such that `first` substituted is equal
-   * to `second`, if one exists. Uses [[matchFormulaRecursive]] as the actual
-   * implementation.
-   *
-   * @param reference the reference formula
-   * @param template the formula to match
-   * @param takenTermVariables any variables in the template which cannot be
-   * substituted, i.e., treated as constant
-   * @param takenFormulaVariables any formula variables in the template which
-   * cannot be substituted, i.e., treated as constant
-   * @return substitution pair (Option) from formula variables to formulas, and
-   * variables to terms. `None` if a substitution does not exist.
-   */
-  def matchFormula(
-      reference: Formula,
-      template: Formula,
-      takenTermVariables: Iterator[VariableLabel] = Iterator.empty,
-      takenFormulaVariables: Iterator[VariableFormulaLabel] = Iterator.empty
-  ): Option[(FormulaSubstitution, TermSubstitution)] = {
-    val context = RewriteContext(
-      takenTermVars = takenTermVariables.toSet,
-      takenFormulaVars = takenFormulaVariables.toSet
-    )
-    matchFormulaRecursive(using context)(reference, template, FormulaSubstitution.empty, TermSubstitution.empty)
-  }
-
-  /**
-   * Implementation for matching formulas. See [[matchFormula]] for the
-   * interface.
-   *
-   * @param context all information about restricted variables and fresh name generation state
-   * @param reference the reference formula
-   * @param template the formula to match
-   * @param formulaSubstitution currently accumulated susbtitutions to formula variables
-   * @param termSubstitution currently accumulated susbtitutions to term variables
-   * @return substitution pair (Option) from formula variables to formulas, and
-   * variables to terms. `None` if a substitution does not exist.
-   */
-  private def matchFormulaRecursive(using
-      context: RewriteContext
-  )(reference: Formula, template: Formula, formulaSubstitution: FormulaSubstitution, termSubstitution: TermSubstitution): Option[(FormulaSubstitution, TermSubstitution)] = {
-    if (isSame(reference, template))
-      Some((formulaSubstitution, termSubstitution))
-    else
-      (reference, template) match {
-        case (BinderFormula(labelR, boundR, innerR), BinderFormula(labelT, boundT, innerT)) if labelR == labelT => {
-          val freshVar = VariableLabel(context.freshIdentifier)
-
-          // add a safety substitution to make sure bound variable isn't substituted, and check instantiated bodies
-          val innerSubst = matchFormulaRecursive(
-            substituteVariables(innerR, Map[VariableLabel, Term](boundR -> freshVar)),
-            substituteVariables(innerT, Map[VariableLabel, Term](boundT -> freshVar)),
-            formulaSubstitution,
-            termSubstitution + (freshVar -> freshVar) // dummy substitution to make sure we don't attempt to match this as a variable
-          )
-
-          innerSubst match {
-            case None => innerSubst
-            case Some((sf, st)) => {
-              val cleanSubst = (sf, st - freshVar) // remove the dummy substitution we added
-
-              // were any formula substitutions involving the bound variable required?
-              // if yes, not matchable
-              if (cleanSubst._1.exists((k, v) => v.freeVariables.contains(freshVar))) None
-              else Some(cleanSubst)
+              if (argCan.exists(_.isEmpty)) None
+              else Some(ConnectorFormula(l1, argCan.map(_.get)))
             }
+            case _ => None
           }
         }
+        case BinderFormula(l1, x1: VariableLabel, inner1) => {
+          second match {
+            case BinderFormula(l2, x2: VariableLabel, inner2) => {
+              val newx = VariableLabel(freshId(takenIds, x1.id))
+              val newInner1 = substituteVariables(inner1, Map[VariableLabel, Term](x1 -> newx))
+              val newInner2 = substituteVariables(inner2, Map[VariableLabel, Term](x2 -> newx))
 
-        case (ConnectorFormula(labelR, argsR), ConnectorFormula(labelT, argsT)) if labelR == labelT =>
-          if (argsR.length != argsT.length)
-            // quick discard
-            None
-          else {
-            // recursively check inner formulas
-            val newSubstitution = (argsR zip argsT).foldLeft(Option(formulaSubstitution, termSubstitution)) {
-              case (Some(substs), (ref, temp)) => matchFormulaRecursive(ref, temp, substs._1, substs._2)
-              case (None, _) => None
+              val innerRes = canReachOneStepTerm2(newInner1, newInner2, subst, takenIds + newx.id)
+
+              if (innerRes.isEmpty) innerRes
+              else Some(BinderFormula(l1, newx, innerRes.get))
             }
-            newSubstitution
+            case _ => None
           }
+        }
+      }
+  }
 
-        case (_, PredicateFormula(labelT: VariableFormulaLabel, _)) =>
-          // can this variable be matched with the reference based on previously known or new substitutions?
-          if (reference == formulaSubstitution.getOrElse(labelT, reference)) Some(formulaSubstitution + (labelT -> reference), termSubstitution)
-          else if (template.label == reference.label && reference == formulaSubstitution.getOrElse(labelT, reference)) Some(formulaSubstitution, termSubstitution)
-          else None
+  /**
+   * Decides a one-step word problem for two given terms and a set of ground
+   * rewrites modulo OL. If possible, returns a context corresponding to the
+   * substitution.
+   *
+   * @param first term
+   * @param second term
+   * @param subst list of possible substitutions
+   * @return
+   */
+  def canReachOneStepTerm(first: Term, second: Term, subst: List[(Term, Term)]): Option[LambdaTermTerm] = {
+    val freeids = (first.freeVariables ++ second.freeVariables).map(_.id)
+    val substWithVar = subst
+      .foldLeft((freeids, Nil: Seq[((Term, Term), Identifier)])) {
+        case ((frs, l), s) => {
+          val x = freshId(frs, "x")
+          (frs + x, l :+ (s, x))
+        }
+      }
+      ._2
+    val body = canReachOneStepTerm2(first, second, substWithVar)
 
-        case (PredicateFormula(labelR, argsR), PredicateFormula(labelT, argsT)) if labelR == labelT =>
-          if (argsR.length != argsT.length)
-            // quick discard
-            None
-          else {
-            // our arguments are terms, match them recursively
-            val newTermSubstitution = (argsR zip argsT).foldLeft(Option(termSubstitution)) {
-              case (Some(tSubst), (ref, temp)) => matchTermRecursive(ref, temp, tSubst)
-              case (None, _) => None
+    if (body.isEmpty) None
+    else Some(lambda(substWithVar.map(s => VariableLabel(s._2)), body.get))
+  }
+
+  /**
+   * Decides a one-step word problem for two given formulas and a set of ground
+   * term rewrites modulo OL. If possible, returns a context corresponding to
+   * the substitution.
+   *
+   * @param first formula
+   * @param second formula
+   * @param subst list of possible substitutions
+   * @return a lambda term -> formula if the word problem is affirmative
+   */
+  def canReachOneStepTermFormula(first: Formula, second: Formula, subst: List[(Term, Term)]): Option[LambdaTermFormula] = {
+    val takenids = (first.freeVariables ++ second.freeVariables).map(_.id)
+    val substWithVar = subst
+      .foldLeft((takenids, Nil: Seq[((Term, Term), Identifier)])) {
+        case ((frs, l), s) => {
+          val x = freshId(frs, "x")
+          (frs + x, l :+ (s, x))
+        }
+      }
+      ._2
+    val body = canReachOneStepTerm2(first, second, substWithVar, takenids ++ substWithVar.map(_._2))
+
+    if (body.isEmpty) None
+    else Some(lambda(substWithVar.map(s => VariableLabel(s._2)), body.get))
+  }
+
+  /**
+   * Internal function for [[canReachOneStepOLFormula]]
+   *
+   * @param first formula
+   * @param second formula
+   * @param subst list of formula substitutions with an assigned identifier
+   * @param takenIds list of taken identifiers for instantiation
+   */
+  def canReachOneStepOLFormula2(first: Formula, second: Formula, subst: Seq[((Formula, Formula), Identifier)], takenIds: Set[Identifier]): Option[Formula] = {
+    lazy val validSubst = subst.find { case ((l, r), _) => isSame(first, l) && isSame(second, r) }
+
+    if (isSame(first, second)) Some(first)
+    else if (validSubst.isDefined) Some(VariableFormulaLabel(validSubst.get._2))
+    else if (first.label != second.label) None
+    else
+      first match {
+        case ConnectorFormula(l1, arg1) => {
+          second match {
+            case ConnectorFormula(l2, arg2) => {
+              val argCan = (arg1 zip arg2).map { case (f, s) => canReachOneStepOLFormula2(f, s, subst, takenIds) }
+
+              if (argCan.exists(_.isEmpty)) None
+              else Some(ConnectorFormula(l1, argCan.map(_.get)))
             }
-            if (newTermSubstitution.isEmpty) None
-            else Some(formulaSubstitution, newTermSubstitution.get)
+            case _ => None
           }
+        }
+        case BinderFormula(l1, x1: VariableLabel, inner1) => {
+          second match {
+            case BinderFormula(l2, x2: VariableLabel, inner2) => {
+              val newx = VariableLabel(freshId(takenIds, x1.id))
+              val newInner1 = substituteVariables(inner1, Map[VariableLabel, Term](x1 -> newx))
+              val newInner2 = substituteVariables(inner2, Map[VariableLabel, Term](x2 -> newx))
+
+              val innerRes = canReachOneStepOLFormula2(newInner1, newInner2, subst, takenIds + newx.id)
+
+              if (innerRes.isEmpty) innerRes
+              else Some(BinderFormula(l2, newx, innerRes.get))
+            }
+            case _ => None
+          }
+        }
         case _ => None
       }
   }
 
-  // rewrites
-
   /**
-   * A term rewrite rule (`l -> r`) with an accompanying instantiation, given
-   * by a term substitution.
+   * Decides a one-step word problem for two given formulas and a set of ground
+   * formula rewrites modulo OL. If possible, returns a context corresponding to
+   * the substitution.
    *
-   * @example A rule without any instantiation would be `((l -> r),
-   * TermSubstitution.empty)`.
-   * @example Commutativity of a function with instantiation can be `((f(x, y)
-   * -> f(y, x)), Map(x -> pair(a, b), y -> c))`
+   * @param first formula
+   * @param second formula
+   * @param subst list of possible formula substitutions
+   * @return a lambda formula -> formula if the word problem is affirmative
    */
-  type TermRule = ((Term, Term), TermSubstitution)
+  def canReachOneStepOLFormula(first: Formula, second: Formula, subst: List[(Formula, Formula)]): Option[LambdaFormulaFormula] = {
+    val takenids = (first.freeVariables ++ second.freeVariables).map(_.id)
+    val substWithVar = subst
+      .foldLeft((takenids, Nil: Seq[((Formula, Formula), Identifier)])) {
+        case ((frs, l), s) => {
+          val x = freshId(frs, "x")
+          (frs + x, l :+ (s, x))
+        }
+      }
+      ._2
+    val body = canReachOneStepOLFormula2(first, second, substWithVar, takenids ++ substWithVar.map(_._2))
 
-  /**
-   * A formula rewrite rule (`l -> r`) with an accompanying instantiation,
-   * given by a formula and a term substitution.
-   *
-   * @example A rule without any instantiation would be `((l -> r),
-   * FormulaSubstitution.empty)`.
-   * @example `((P(x) \/ Q -> Q /\ R(x)), Map(Q -> A \/ B, x -> f(t)))`
-   */
-  type FormulaRule = ((Formula, Formula), (FormulaSubstitution, TermSubstitution))
-
-  /**
-   * A lambda representing a term, with inputs as terms. Carries extra
-   * information about rewrite rules used in its construction for proof
-   * geenration later.
-   *
-   * @param termVars variables in the body to be treated as parameters closed
-   * under this function
-   * @param termRules mapping to the rules (with instantiations) used to
-   * construct this function; used for proof construction
-   * @param body the body of the function
-   */
-  case class TermRewriteLambda(
-      termVars: Seq[VariableLabel] = Seq.empty,
-      termRules: Seq[(VariableLabel, TermRule)] = Seq.empty,
-      body: Term
-  ) {}
-
-  /**
-   * A lambda representing a formula, with inputs as terms or formulas. Carries
-   * extra information about rewrite rules used in its construction for proof
-   * geenration later.
-   *
-   * @param termVars variables in the body to be treated as parameters closed
-   * under this function
-   * @param formulaVars formula variables in the body to be treated as
-   * parameters closed under this function
-   * @param termRules mapping to the term rewrite rules (with instantiations)
-   * used to construct this function; used for proof construction
-   * @param formulaRules mapping to the formula rewrite rules (with
-   * instantiations) used to construct this function; used for proof
-   * construction
-   * @param body the body of the function
-   */
-  case class FormulaRewriteLambda(
-      termRules: Seq[(VariableLabel, TermRule)] = Seq.empty,
-      formulaRules: Seq[(VariableFormulaLabel, FormulaRule)] = Seq.empty,
-      body: Formula
-  ) {}
-
-  /**
-   * Dummy connector used to combine formulas for convenience during rewriting
-   */
-  val formulaRewriteConnector = SchematicConnectorLabel(Identifier("@@rewritesTo@@"), 2)
-
-  /**
-   * Dummy function symbol used to combine terms for convenience during rewriting
-   */
-  val termRewriteConnector = ConstantFunctionLabel(Identifier("@@rewritesTo@@"), 2)
-
-  /**
-   * Decides whether a term `first` be rewritten into `second` at the top level
-   * using the provided rewrite rule (with instantiation).
-   *
-   * Reduces to matching using [[matchTermRecursive]].
-   */
-  private def canRewrite(using context: RewriteContext)(first: Term, second: Term, rule: (Term, Term)): Option[TermSubstitution] =
-    matchTermRecursive(termRewriteConnector(first, second), termRewriteConnector(rule._1, rule._2), TermSubstitution.empty)
-
-  /**
-   * Decides whether a formula `first` be rewritten into `second` at the top
-   * level using the provided rewrite rule (with instantiation). Produces the
-   * instantiation as output, if one exists.
-   *
-   * Reduces to matching using [[matchFormulaRecursive]].
-   */
-  private def canRewrite(using context: RewriteContext)(first: Formula, second: Formula, rule: (Formula, Formula)): Option[(FormulaSubstitution, TermSubstitution)] =
-    matchFormulaRecursive(formulaRewriteConnector(first, second), formulaRewriteConnector(rule._1, rule._2), FormulaSubstitution.empty, TermSubstitution.empty)
-
-  /**
-   * Decides whether a term `first` can be rewritten into another term `second`
-   * under the given rewrite rules and restrictions.
-   *
-   * Calls [[getContextRecursive]] as its actual implementation.
-   *
-   * @param first source term
-   * @param second destination term
-   * @param freeTermRules rewrite rules with unrestricted instantiations
-   * @param confinedTermRules rewrite rules with restricted instantiations wrt takenTermVariables
-   * @param takenTermVariables variables to *not* instantiate, i.e., treat as constant, for confined rules
-   */
-  def getContextTerm(
-      first: Term,
-      second: Term,
-      freeTermRules: Seq[(Term, Term)],
-      confinedTermRules: Seq[(Term, Term)] = Seq.empty,
-      takenTermVariables: Set[VariableLabel] = Set.empty
-  ): Option[TermRewriteLambda] = {
-    val context = RewriteContext(
-      takenTermVars = takenTermVariables,
-      freeTermRules = freeTermRules,
-      confinedTermRules = confinedTermRules
-    )
-    getContextRecursive(using context)(first, second)
+    if (body.isEmpty) None
+    else Some(lambda(substWithVar.map(s => VariableFormulaLabel(s._2)), body.get))
   }
 
   /**
-   * Inner implementation for [[getContextTerm]].
+   * Performs a single step parallel rewrite on a term from given ground term
+   * substitutions
    *
-   * @param context all information about rewrite rules and allowed instantiations
-   * @param first source term
-   * @param second destination term
+   * @param first term to substitute in
+   * @param subst list containing pairs representing rewrite rules (l -> r)
    */
-  private def getContextRecursive(using context: RewriteContext)(first: Term, second: Term): Option[TermRewriteLambda] = {
-    // check if there exists a substitution
-    lazy val validSubstitution =
-      context.confinedTermRules
-        .getFirst { case (l, r) =>
-          val subst = canRewrite(using context)(first, second, (l, r))
-          subst.map(s => ((l, r), s))
+  def rewriteOneStepTerm(first: Term, subst: List[(Term, Term)]): Term = {
+    val foundSubst = subst.find { case (l, r) => isSameTerm(first, l) }
+
+    if (foundSubst.isDefined) foundSubst.get._2
+    else Term(first.label, first.args.map(rewriteOneStepTerm(_, subst)))
+  }
+
+  /**
+   * Performs a single step parallel rewrite on a formula from given ground term
+   * substitutions
+   *
+   * @param first formula to substitute in
+   * @param subst list containing pairs representing rewrite rules (l -> r)
+   */
+  def rewriteOneStepTermInFormula(first: Formula, subst: List[(Term, Term)], freeVars: Option[Set[Identifier]] = None): Formula = {
+    val freeVarsInner =
+      if (freeVars.isDefined) freeVars else Some((first.freeVariables ++ subst.foldLeft(Set[VariableLabel]()) { case (frs, (k, v)) => frs ++ k.freeVariables ++ v.freeVariables }).map(_.id))
+    first match {
+      case PredicateFormula(l, arg) => PredicateFormula(l, arg.map(rewriteOneStepTerm(_, subst)))
+      case ConnectorFormula(l, arg) => ConnectorFormula(l, arg.map(rewriteOneStepTermInFormula(_, subst, freeVarsInner)))
+      case BinderFormula(l, x: VariableLabel, inner) => {
+        val newx = VariableLabel(freshId(freeVarsInner.get, x.id))
+        val newInner = substituteVariables(inner, Map[VariableLabel, Term](x -> newx))
+
+        BinderFormula(l, newx, rewriteOneStepTermInFormula(newInner, subst, Some(freeVarsInner.get + newx.id)))
+      }
+    }
+  }
+
+  /**
+   * Performs a single step parallel rewrite on a formula from given ground
+   * formula substitutions
+   *
+   * @param first formula to substitute in
+   * @param subst list containing pairs representing rewrite rules (l -> r)
+   */
+  def rewriteOneStepOLFormulaInFormula(first: Formula, subst: List[(Formula, Formula)], freeVars: Option[Set[Identifier]] = None): Formula = {
+    lazy val freeVarsInner =
+      if (freeVars.isDefined) freeVars else Some((first.freeVariables ++ subst.foldLeft(Set[VariableLabel]()) { case (frs, (k, v)) => frs ++ k.freeVariables ++ v.freeVariables }).map(_.id))
+    val foundSubst = subst.find { case (k, v) => isSame(first, k) }
+
+    if (foundSubst.isDefined) foundSubst.get._2
+    else
+      first match {
+        case PredicateFormula(_, _) => first
+        case ConnectorFormula(l, arg) => ConnectorFormula(l, arg.map(rewriteOneStepOLFormulaInFormula(_, subst, freeVarsInner)))
+        case BinderFormula(l, x: VariableLabel, inner) => {
+          val newx = VariableLabel(freshId(freeVarsInner.get, x.id))
+          val newInner = substituteVariables(inner, Map[VariableLabel, Term](x -> newx))
+
+          BinderFormula(l, newx, rewriteOneStepOLFormulaInFormula(newInner, subst, Some(freeVarsInner.get + newx.id)))
         }
-        .orElse {
-          // free all variables for substitution
-          // matchTermRecursive does not generate any free variables
-          // so it cannot affect global state, so this is safe to do
-          val freeContext = context.copy(takenTermVars = Set.empty)
-          freeContext.freeTermRules.getFirst { case (l, r) =>
-            val subst = canRewrite(using freeContext)(first, second, (l, r))
-            subst.map(s => ((l, r), s))
+      }
+  }
+
+  def getContextOneStepTerm2(first: Term, subst: List[((Term, Term), Identifier)]): Term = {
+    val validSubst = subst.find { case ((l, _), _) => isSameTerm(first, l) }
+
+    if (validSubst.isDefined) VariableLabel(validSubst.get._2)
+    else Term(first.label, first.args.map(getContextOneStepTerm2(_, subst)))
+  }
+
+  def getContextOneStepFormula2(first: Formula, formSubst: List[((Formula, Formula), Identifier)], termSubst: List[((Term, Term), Identifier)], takenIds: Set[Identifier]): Formula = {
+    lazy val validSubst = formSubst.find { case ((l, _), _) => isSame(first, l) }
+
+    if (validSubst.isDefined) VariableFormulaLabel(validSubst.get._2)
+    else
+      first match {
+        case ConnectorFormula(l1, arg1) => ConnectorFormula(l1, arg1.map(getContextOneStepFormula2(_, formSubst, termSubst, takenIds)))
+        case BinderFormula(l1, x1: VariableLabel, inner1) => {
+          val newx = VariableLabel(freshId(takenIds, x1.id))
+          val newInner1 = substituteVariables(inner1, Map[VariableLabel, Term](x1 -> newx))
+
+          BinderFormula(l1, newx, getContextOneStepFormula2(newInner1, formSubst, termSubst, takenIds + newx.id))
+        }
+        case PredicateFormula(l1, arg1) => PredicateFormula(l1, arg1.map(getContextOneStepTerm2(_, termSubst)))
+      }
+  }
+
+  def getContextOneStepFormula(first: Formula, formSubst: List[(Formula, Formula)], termSubst: List[(Term, Term)]): LambdaTermFormulaFormula = {
+    val takenids = first.freeVariables.map(_.id)
+    val formSubstWithVar = formSubst
+      .foldLeft((takenids, Nil: List[((Formula, Formula), Identifier)])) {
+        case ((frs, l), s) => {
+          val x = freshId(frs, "x")
+          (frs + x, l :+ (s, x))
+        }
+      }
+    val termSubstWithVar = termSubst
+      .foldLeft((formSubstWithVar._1, Nil: List[((Term, Term), Identifier)])) {
+        case ((frs, l), s) => {
+          val x = freshId(frs, "x")
+          (frs + x, l :+ (s, x))
+        }
+      }
+
+    val body = getContextOneStepFormula2(first, formSubstWithVar._2, termSubstWithVar._2, termSubstWithVar._1)
+
+    (
+      formSubstWithVar._2.map(s => VariableFormulaLabel(s._2)),
+      termSubstWithVar._2.map(s => VariableLabel(s._2)),
+      body
+    )
+  }
+
+  /**
+   * Extension methods for rewrites
+   */
+  extension (t: Term) {
+    def substituted(subst: (Term, Term)*): Term = rewriteOneStepTerm(t, subst.toList)
+  }
+
+  extension (f: Formula) {
+    def substitutedTerms(subst: (Term, Term)*): Formula = rewriteOneStepTermInFormula(f, subst.toList)
+
+    def substituted(subst: (Formula, Formula)*): Formula = rewriteOneStepOLFormulaInFormula(f, subst.toList)
+  }
+
+  def canReachOneStep2(first: Formula, second: Formula, formSubst: List[((Formula, Formula), Identifier)], termSubst: List[((Term, Term), Identifier)], takenIds: Set[Identifier]): Option[Formula] = {
+    lazy val validSubst = formSubst.find { case ((l, r), _) => isSame(first, l) && isSame(second, r) }
+
+    if (isSame(first, second)) Some(first)
+    else if (validSubst.isDefined) Some(VariableFormulaLabel(validSubst.get._2))
+    else if (first.label != second.label) None
+    else
+      first match {
+        case ConnectorFormula(l1, arg1) => {
+          second match {
+            case ConnectorFormula(l2, arg2) => {
+              val argCan = (arg1 zip arg2).map { case (f, s) => canReachOneStep2(f, s, formSubst, termSubst, takenIds) }
+
+              if (argCan.exists(_.isEmpty)) None
+              else Some(ConnectorFormula(l1, argCan.map(_.get)))
+            }
+            case _ => None
           }
         }
+        case BinderFormula(l1, x1: VariableLabel, inner1) => {
+          second match {
+            case BinderFormula(l2, x2: VariableLabel, inner2) => {
+              val newx = VariableLabel(freshId(takenIds, x1.id))
+              val newInner1 = substituteVariables(inner1, Map[VariableLabel, Term](x1 -> newx))
+              val newInner2 = substituteVariables(inner2, Map[VariableLabel, Term](x2 -> newx))
 
-    if (first == second) Some(TermRewriteLambda(body = first))
-    else if (validSubstitution.isDefined) {
-      val newVar = VariableLabel(context.freshIdentifier)
-      val body = Term(newVar, Seq.empty) // newVar()
+              val innerRes = canReachOneStep2(newInner1, newInner2, formSubst, termSubst, takenIds + newx.id)
+
+              if (innerRes.isEmpty) innerRes
+              else Some(BinderFormula(l1, newx, innerRes.get))
+            }
+            case _ => None
+          }
+        }
+        case PredicateFormula(l1, arg1) => {
+          second match {
+            case PredicateFormula(l2, arg2) => {
+              val argCan = (arg1 zip arg2).map { case (f, s) => canReachOneStepTerm2(f, s, termSubst) }
+
+              if (argCan.exists(_.isEmpty)) None
+              else Some(PredicateFormula(l1, argCan.map(_.get)))
+            }
+            case _ => None
+          }
+        }
+      }
+  }
+
+  def canReachOneStep(first: Formula, second: Formula, formSubst: List[(Formula, Formula)], termSubst: List[(Term, Term)]): Option[LambdaTermFormulaFormula] = {
+    val takenids = (first.freeVariables ++ second.freeVariables).map(_.id)
+    val formSubstWithVar = formSubst
+      .foldLeft((takenids, Nil: List[((Formula, Formula), Identifier)])) {
+        case ((frs, l), s) => {
+          val x = freshId(frs, "x")
+          (frs + x, l :+ (s, x))
+        }
+      }
+    val termSubstWithVar = termSubst
+      .foldLeft((formSubstWithVar._1, Nil: List[((Term, Term), Identifier)])) {
+        case ((frs, l), s) => {
+          val x = freshId(frs, "x")
+          (frs + x, l :+ (s, x))
+        }
+      }
+
+    val body = canReachOneStep2(first, second, formSubstWithVar._2, termSubstWithVar._2, termSubstWithVar._1)
+
+    if (body.isEmpty) None
+    else
       Some(
-        TermRewriteLambda(
-          Seq(newVar),
-          Seq(newVar -> validSubstitution.get),
-          body
+        (
+          formSubstWithVar._2.map(s => VariableFormulaLabel(s._2)),
+          termSubstWithVar._2.map(s => VariableLabel(s._2)),
+          body.get
         )
       )
-    } else if (first.label != second.label || first.args.length != second.args.length) None
-    else {
-      // recurse
-      // known: first.label == second.label
-      // first.args.lnegth == second.args.length
-      // and first cannot be rewritten into second
-      val innerSubstitutions = (first.args zip second.args).map(arg => getContextRecursive(using context)(arg._1, arg._2))
-
-      if (innerSubstitutions.exists(_.isEmpty)) None
-      else {
-        val retrieved = innerSubstitutions.map(_.get)
-        val body = Term(first.label, retrieved.map(_.body))
-        val lambda =
-          retrieved.foldLeft(TermRewriteLambda(body = body)) { case (currentLambda, nextLambda) =>
-            TermRewriteLambda(
-              currentLambda.termVars ++ nextLambda.termVars,
-              currentLambda.termRules ++ nextLambda.termRules,
-              currentLambda.body
-            )
-          }
-        Some(lambda)
-      }
-    }
   }
 
-  /**
-   * Decides whether a formula `first` can be rewritten into another formula
-   * `second` under the given rewrite rules and restrictions.
-   *
-   * Calls [[getContextRecursive]] as its actual implementation.
-   *
-   * @param first source formula
-   * @param second destination formula
-   * @param freeTermRules term rewrite rules with unrestricted instantiations
-   * @param freeFormulaRules formula rewrite rules with unrestricted
-   * instantiations
-   * @param confinedTermRules term rewrite rules with restricted instantiations
-   * wrt takenTermVariables
-   * @param confinedTermRules formula rewrite rules with restricted
-   * instantiations wrt takenTermVariables
-   * @param takenTermVariables term variables to *not* instantiate, i.e., treat
-   * as constant, for confined rules
-   * @param takenFormulaVariables formula variables to *not* instantiate, i.e.,
-   * treat as constant, for confined rules
-   */
-  def getContextFormula(
-      first: Formula,
-      second: Formula,
-      freeTermRules: Seq[(Term, Term)],
-      freeFormulaRules: Seq[(Formula, Formula)],
-      confinedTermRules: Seq[(Term, Term)] = Seq.empty,
-      takenTermVariables: Set[VariableLabel] = Set.empty,
-      confinedFormulaRules: Seq[(Formula, Formula)] = Seq.empty,
-      takenFormulaVariables: Set[VariableFormulaLabel] = Set.empty
-  ): Option[FormulaRewriteLambda] = {
-    val context = RewriteContext(
-      takenTermVars = takenTermVariables,
-      takenFormulaVars = takenFormulaVariables,
-      freeTermRules = freeTermRules,
-      confinedTermRules = confinedTermRules,
-      freeFormulaRules = freeFormulaRules,
-      confinedFormulaRules = confinedFormulaRules
-    )
-    getContextRecursive(using context)(first, second)
-  }
+  // case class RewriteSet(
+  //   freeFormulaRules: Seq[(Formula, Formula)],
+  //   freeTermRules: Seq[(Term, Term)],
+  //   confinedFormulaRules: Seq[(Formula, Formula)],
+  //   confinedTermRules: Seq[(Term, Term)],
+  //   takenFormulaVars: Seq[VariableFormulaLabel],
+  //   takenTermVars: Seq[VariableLabel],
+  //   lastID: Identifier = freshId((takenFormulaVars ++ takenTermVars), "__rewriteVar__")
+  // ){
+  //   def freshIdentifier = {
+  //     val newID = freshId(Seq(lastID), "__rewriteVar__")
+  //     (newID, this.copy(lastID = newID))
+  //   }
+  // }
 
-  def getContextFormulaSet(
-      first: Seq[Formula],
-      second: Seq[Formula],
-      freeTermRules: Seq[(Term, Term)],
-      freeFormulaRules: Seq[(Formula, Formula)],
-      confinedTermRules: Seq[(Term, Term)] = Seq.empty,
-      takenTermVariables: Set[VariableLabel] = Set.empty,
-      confinedFormulaRules: Seq[(Formula, Formula)] = Seq.empty,
-      takenFormulaVariables: Set[VariableFormulaLabel] = Set.empty
-  ): Option[Seq[FormulaRewriteLambda]] = {
-    val context = RewriteContext(
-      takenTermVars = takenTermVariables,
-      takenFormulaVars = takenFormulaVariables,
-      freeTermRules = freeTermRules,
-      confinedTermRules = confinedTermRules,
-      freeFormulaRules = freeFormulaRules,
-      confinedFormulaRules = confinedFormulaRules
-    )
+  // type FormulaSubstitution = Map[VariableFormulaLabel, Formula]
+  // type TermSubstitution = Map[VariableLabel, Term]
 
-    val substSeq = first.map { f =>
-      second.getFirst { s =>
-        val newContext = context.copy()
-        val subst = getContextRecursive(using newContext)(f, s)
-        subst.foreach { _ => context.updateTo(newContext) }
-        subst
-      }
-    }
+  // case class RewriteContext(
+  //   // the variable, the rule it represents, and with what instantiations
+  //   formulaVars: Seq[(VariableFormulaLabel, (Formula, Formula), FormulaSubstitution)],
+  //   termVars: Seq[(VariableLabel, (Term, Term), TermSubstitution)],
+  //   body: Formula
+  // )
 
-    // Seq[Option[_]] -> Option[Seq[_]]
-    substSeq.foldLeft(Option(Seq.empty[FormulaRewriteLambda]))((f, s) => f.flatMap(f1 => s.map(s1 => f1 :+ s1)))
-  }
+  // def canRewriteList(l: Seq[(Formula, Formula)], ruleSet: RewriteSet) =
+  //     l.fold(Some(Seq[RewriteContext](), ruleSet)) {
+  //       case (Some(seq, rSet), (f, s)) =>
+  //         val can = canRewrite(f, s, rSet)
+  //         can match {
+  //           case Some((c, r)) => (seq :+ c, r)
+  //           case None => None
+  //         }
+  //       case None => None
+  //     }
 
-  /**
-   * Inner implementation for [[getContextFormula]].
-   *
-   * @param context all information about rewrite rules and allowed instantiations
-   * @param first source formula
-   * @param second destination formula
-   */
-  private def getContextRecursive(using context: RewriteContext)(first: Formula, second: Formula): Option[FormulaRewriteLambda] = {
-    // check if there exists a substitution
-    lazy val validSubstitution =
-      context.confinedFormulaRules
-        .getFirst { (l: Formula, r: Formula) =>
-          val subst = canRewrite(using context)(first, second, (l, r))
-          subst.map(s => ((l, r), s))
-        }
-        .orElse {
-          // free all variables for substitution
-          // matchFormulaRecursive generates but does not expose any new variables
-          // It cannot affect global state, so this is safe to do
-          val freeContext = context.copy(takenTermVars = Set.empty)
-          freeContext.freeFormulaRules.getFirst { case (l, r) =>
-            val subst = canRewrite(using freeContext)(first, second, (l, r))
-            subst.map(s => ((l, r), s))
-          }
-        }
+  // def canRewriteTermList(l: Seq[(Term, Term)], ruleSet: RewriteSet): Option[(RewriteContext)] =
+  //     l.fold(Some(Seq[RewriteContext](), ruleSet)) {
+  //       case (Some(seq, rSet), (f, s)) =>
+  //         val can = canRewrite(f, s, rSet)
+  //         can match {
+  //           case Some((c, r)) => (seq :+ c, r)
+  //           case None => None
+  //         }
+  //       case None => None
+  //     }
 
-    if (isSame(first, second)) Some(FormulaRewriteLambda(body = first))
-    else if (validSubstitution.isDefined) {
-      val newVar = VariableFormulaLabel(context.freshIdentifier)
-      val body = PredicateFormula(newVar, Seq.empty) // newVar()
-      Some(
-        FormulaRewriteLambda(
-          Seq(),
-          Seq(newVar -> validSubstitution.get),
-          body
-        )
-      )
-    } else if (first.label != second.label)
-      None
-    else {
-      // recurse
-      // known: first.label == second.label
-      // and first cannot be rewritten into second
-      (first, second) match {
-        case (BinderFormula(labelF, boundF, innerF), BinderFormula(labelS, boundS, innerS)) => {
-          val freshVar = VariableLabel(context.freshIdentifier)
-          val freeContext = context.copy(takenTermVars = context.takenTermVars + freshVar)
+  // case object RewriteConnector extends ConstantConnectorLabel(Identifier("__rewritesTo__"), 2)
 
-          // add a safety substitution to make sure bound variable isn't substituted, and check instantiated bodies
-          val innerSubst = getContextRecursive(using freeContext)(
-            substituteVariables(innerF, Map[VariableLabel, Term](boundF -> freshVar)),
-            substituteVariables(innerS, Map[VariableLabel, Term](boundS -> freshVar))
-          )
+  // def canRewrite(first: Formula, second: Formula, ruleSet: RewriteSet): Option[(RewriteContext, RewriteSet)] = {
+  //   lazy val takenVars = first.freeVariables ++ first.freeVariableFormulaLabels ++ second.freeVariables ++ second.freeVariableFormulaLabels
+  //   lazy val replaceableVars = takenVars -- ruleSet.takenFormulaVars -- ruleSet.takenTermVars
+  //   lazy val validSubst =
+  //     ruleSet.confinedFormulaRules.collectFirst {
+  //       case (l, r) =>
+  //         val subst = matchFormula(RewriteConnector(l, r), RewriteConnector(first, second), vars = Some(replaceableVars))
+  //         subst match {
+  //           case Some(s) => s
+  //         }
+  //     }
+  //     .orElse (
+  //       ruleSet.freeFormulaRules.collectFirst {
+  //       case (l, r) =>
+  //         val subst = matchFormula(RewriteConnector(l, r), RewriteConnector(first, second), vars = None)
+  //         subst match {
+  //           case Some(s) => s
+  //         }
+  //       }
+  //     )
 
-          context.updateTo(freeContext)
+  //   if (isSame(first, second))
+  //     // done
+  //     Some((RewriteContext(Seq(), Seq(), first), ruleSet))
+  //   else if (validSubst.isDefined)
+  //     RewriteContext(
 
-          innerSubst.map(s => s.copy(body = BinderFormula(labelF, freshVar, s.body)))
-        }
+  //     )
+  //   else if (first.label != second.label)
+  //     None
+  //   else
+  //     // recurse
+  //     first match {
+  //       case ConnectorFormula(l1, arg1) => {
+  //         second match {
+  //           case ConnectorFormula(l2, arg2) => {
+  //             val argCan = canRewriteList(arg1 zip arg2)
 
-        case (ConnectorFormula(labelF, argsF), ConnectorFormula(labelS, argsS)) =>
-          if (argsF.length != argsS.length)
-            // quick discard
-            None
-          else {
-            // recursively check inner formulas
-            val innerSubstitutions = (argsF zip argsS).map(arg => getContextRecursive(using context)(arg._1, arg._2))
+  //             if (argCan.isEmpty) None
+  //             else
+  //               val newArgs = argCan.get._1
+  //               val newRuleSet = argCan.get._2
+  //               val intermediateContext = newArgs.reduce {
+  //                 // combine the generated substitutions:
+  //                 // By construction, there are no variable conflicts, so we
+  //                 // can combine them blindly.
+  //                 // This can generate redundant variables for the same rewrite
+  //                 // rule, but the added proof checking is a lot less work than
+  //                 // checking for this and appropriately combining the
+  //                 // variables.
+  //                 case (f, s) =>
+  //                   RewriteContext(
+  //                     formulaVars = f.formulaVars ++ s.formulaVars,
+  //                     termVars = f.termVars ++ s.termVars,
+  //                     body = top()
+  //                   )
+  //               }
+  //               val newBody = ConnectorFormula(l1, newArgs.map(_.body))
+  //               Some(
+  //                 (
+  //                   intermediateContext.copy(body = newBody),
+  //                   newRuleSet
+  //                 )
+  //               )
+  //           }
+  //           case _ => None
+  //         }
+  //       }
+  //       case BinderFormula(l1, x1: VariableLabel, inner1) => {
+  //         second match {
+  //           case BinderFormula(l2, x2: VariableLabel, inner2) => {
+  //             val newID, newRuleSet = ruleSet.freshIdentifier
+  //             val newX = VariableLabel(newID)
+  //             val newInner1 = substituteVariables(inner1, Map[VariableLabel, Term](x1 -> newX))
+  //             val newInner2 = substituteVariables(inner2, Map[VariableLabel, Term](x2 -> newX))
 
-            if (innerSubstitutions.exists(_.isEmpty)) None
-            else {
-              val retrieved = innerSubstitutions.map(_.get)
-              val body = ConnectorFormula(labelF, retrieved.map(_.body))
-              val lambda =
-                retrieved.foldLeft(FormulaRewriteLambda(body = body)) { case (currentLambda, nextLambda) =>
-                  FormulaRewriteLambda(
-                    currentLambda.termRules ++ nextLambda.termRules,
-                    currentLambda.formulaRules ++ nextLambda.formulaRules,
-                    currentLambda.body
-                  )
-                }
-              Some(lambda)
-            }
-          }
+  //             val innerRes = canRewrite(newInner1, newInner2, newRuleSet)
 
-        case (PredicateFormula(labelF, argsF), PredicateFormula(labelS, argsS)) =>
-          if (argsF.length != argsS.length)
-            // quick discard
-            None
-          else {
-            // our arguments are terms, get contexts from them recursively
-            val innerSubstitutions = (argsF zip argsS).map(arg => getContextRecursive(using context)(arg._1, arg._2))
+  //             if (innerRes.isEmpty) None
+  //             else
+  //               val newInner = innerRes.get._1
+  //               val newRuleSet = innerRes.get._2
+  //               Some(
+  //                 (
+  //                   newInner.copy(body = BinderFormula(l1, newX, innerRes.body)),
+  //                   newRuleSet
+  //                 )
+  //               )
+  //           }
+  //           case _ => None
+  //         }
+  //       }
+  //       case PredicateFormula(l1, arg1) => {
+  //         second match {
+  //           case PredicateFormula(l2, arg2) => {
+  //             val argCan = canRewriteTermList(arg1 zip arg2, ruleSet)
 
-            if (innerSubstitutions.exists(_.isEmpty)) None
-            else {
-              val retrieved = innerSubstitutions.map(_.get)
-              val body = PredicateFormula(labelF, retrieved.map(_.body))
-              val lambda =
-                retrieved.foldLeft(FormulaRewriteLambda(body = body)) { case (currentLambda, nextLambda) =>
-                  FormulaRewriteLambda(
-                    currentLambda.termRules ++ nextLambda.termRules,
-                    currentLambda.formulaRules,
-                    currentLambda.body
-                  )
-                }
-              Some(lambda)
-            }
-          }
-        case _ => None
-      }
-    }
-  }
+  //             if (argCan.isEmpty) None
+  //             else
+  //               val newArgs = argCan.get._1
+  //               val newRuleSet = argCan.get._2
+  //               val intermediateContext = newArgs.reduce {
+  //                 // combine the generated substitutions:
+  //                 // By construction, there are no variable conflicts, so we
+  //                 // can combine them blindly.
+  //                 // This can generate redundant variables for the same rewrite
+  //                 // rule, but the added proof checking is a lot less work than
+  //                 // checking for this and appropriately combining the
+  //                 // variables.
+  //                 case (f, s) =>
+  //                   RewriteContext(
+  //                     formulaVars = f.formulaVars ++ s.formulaVars,
+  //                     termVars = f.termVars ++ s.termVars,
+  //                     body = top()
+  //                   )
+  //               }
+  //               val newBody = ConnectorFormula(l1, newArgs.map(_.body))
+  //               Some(
+  //                 (
+  //                   intermediateContext.copy(body = newBody),
+  //                   newRuleSet
+  //                 )
+  //               )
+  //           }
+  //           case _ => None
+  //         }
+  //       }
+  //     }
+  // }
 }
