+++ src/main/scala/lisa/automation/kernel/SimpleSimplifier.scala
@@ -9,7 +9,7 @@ import lisa.prooflib.ProofTacticLib.*
 import lisa.prooflib.SimpleDeducedSteps
 import lisa.utils.FOLPrinter
 import lisa.utils.KernelHelpers.{_, given}
-import lisa.utils.unification.UnificationUtils
+import lisa.utils.unification.UnificationUtils2
 
 import scala.annotation.nowarn
 import scala.annotation.tailrec
@@ -274,7 +274,7 @@ object SimpleSimplifier {
       val equalities = bot.left.collect { case PredicateFormula(`equality`, Seq(l, r)) =>
         (l, r)
       }
-      val canReach = UnificationUtils.canReachOneStepTermFormula(premRight, botRight, equalities.toList)
+      val canReach = UnificationUtils2.canReachOneStepTermFormula(premRight, botRight, equalities.toList)
 
       if (canReach.isEmpty) {
         proof.InvalidProofTactic("Could not find a set of equalities to rewrite premise into conclusion successfully.")
@@ -319,10 +319,10 @@ object SimpleSimplifier {
           }
       }
 
-      val canReach = UnificationUtils.canReachOneStepTermFormula(premRight, botRight, equalities.toList)
+      val canReach = UnificationUtils2.canReachOneStepTermFormula(premRight, botRight, equalities.toList)
 
       if (canReach.isEmpty || equalities.isEmpty) {
-        val canReach2 = UnificationUtils.canReachOneStepOLFormula(premRight, botRight, iffs.toList)
+        val canReach2 = UnificationUtils2.canReachOneStepOLFormula(premRight, botRight, iffs.toList)
         if (canReach2.isEmpty) {
           proof.InvalidProofTactic("Iffs and Equalities failed")
         } else {
@@ -409,8 +409,8 @@ object SimpleSimplifier {
         case _ => true
       }).toSeq
 
-      lazy val rightPairs = premiseSequent.right zip premiseSequent.right.map(x => bot.right.find(y => UnificationUtils.canReachOneStep(x, y, iffs, eqs).isDefined))
-      lazy val leftPairs = filteredPrem zip filteredPrem.map(x => filteredBot.find(y => UnificationUtils.canReachOneStep(x, y, iffs, eqs).isDefined))
+      lazy val rightPairs = premiseSequent.right zip premiseSequent.right.map(x => bot.right.find(y => UnificationUtils2.canReachOneStep(x, y, iffs, eqs).isDefined))
+      lazy val leftPairs = filteredPrem zip filteredPrem.map(x => filteredBot.find(y => UnificationUtils2.canReachOneStep(x, y, iffs, eqs).isDefined))
 
       lazy val violatingFormulaLeft = leftPairs.find(_._2.isEmpty)
       lazy val violatingFormulaRight = rightPairs.find(_._2.isEmpty)
@@ -422,8 +422,8 @@ object SimpleSimplifier {
       else {
         // actually construct proof
         try {
-          val leftLambdas = leftPairs.collect { case (l, Some(r)) => UnificationUtils.canReachOneStep(l, r, iffs, eqs).get }
-          val rightLambdas = rightPairs.collect { case (l, Some(r)) => UnificationUtils.canReachOneStep(l, r, iffs, eqs).get }
+          val leftLambdas = leftPairs.collect { case (l, Some(r)) => UnificationUtils2.canReachOneStep(l, r, iffs, eqs).get }
+          val rightLambdas = rightPairs.collect { case (l, Some(r)) => UnificationUtils2.canReachOneStep(l, r, iffs, eqs).get }
 
           val eqsForm = eqs.map { case (l, r) => PredicateFormula(`equality`, Seq(l, r)) }.toSet
           val iffsForm = iffs.map { case (l, r) => ConnectorFormula(Iff, Seq(l, r)) }.toSet
@@ -565,8 +565,8 @@ object SimpleSimplifier {
         case _ => true
       }
 
-      lazy val rightLambdas = premiseSequent.right.map(UnificationUtils.getContextOneStepFormula(_, iffs, eqs))
-      lazy val leftLambdas = filteredPrem.map(UnificationUtils.getContextOneStepFormula(_, iffs, eqs))
+      lazy val rightLambdas = premiseSequent.right.map(UnificationUtils2.getContextOneStepFormula(_, iffs, eqs))
+      lazy val leftLambdas = filteredPrem.map(UnificationUtils2.getContextOneStepFormula(_, iffs, eqs))
 
       // actually construct proof
       val eqsForm = eqs.map { case (l, r) => PredicateFormula(`equality`, Seq(l, r)) }.toSet
